LN-283 Fix the delete button on draft ballot

+++ src/www.lidonation.com/var/www/resources/js/catalyst-explorer/modules/bookmarks/DraftBallotGroupCard.vue
@@ -191,8 +191,10 @@ function removeItem(id: number) {
     if (canDelete) {
         axios.delete(route('catalystExplorer.bookmarkItem.delete', { bookmarkItem: id }))
             .then((res) => {
-                bookmarksStore.deleteItem(id, collectionHash.value);
-                bookmarksStore.loadDraftBallot();
+                if(res.status === 200){
+                    bookmarksStore.deleteItem(id, collectionHash.value);
+                    bookmarksStore.loadDraftBallot();
+                }
             })
             .catch((error) => {
                 if (error.response && error.response.status === 403) {
